Add AI Model Cataloger based on config.json (#24)
Adds a new cataloger to parse the config.json file associated with AI Models, along with model card.
Detects AI Model weight found in a Container image or a directory.
For a Model card, config.json or a repository, use its remote host to get the unique model name
For a valid Model card, use the respective AI Model package with base model as a dependency
A Model file inside a container that could be resolved will have "huggingface" as its namespace
A Model file that could not be resolved will have "custommodel" as its namespace